+++ src/Ext/Sidebar/Hooks.cpp
@@ -97,3 +97,30 @@ DEFINE_HOOK(0x72FCB5, InitSideRectangles_CenterBackground, 0x5)
 
 	return 0;
 }
+
+DEFINE_HOOK(0x6A557A, SidebarClass_Init_IO_RecordDiplomacyHouses1, 0x5)
+{
+	enum { SkipGameCode = 0x6A5830, ContinueGameCode = 0x6A558D };
+
+	const GameMode mode = SessionClass::Instance->GameMode;
+
+	return (mode == GameMode::Skirmish || mode == GameMode::LAN || mode == GameMode::Internet) ? ContinueGameCode : SkipGameCode;
+}
+
+DEFINE_HOOK(0x6A55BF, SidebarClass_Init_IO_RecordDiplomacyHouses2, 0x7)
+{
+	enum { ContinueLoop = 0x6A55CF, BreakLoop = 0x6A55C8 };
+
+	GET(HouseClass*, pHouse, EAX);
+
+	return (pHouse->IsHumanPlayer || HouseClass::CurrentPlayer == HouseClass::Observer) ? BreakLoop : ContinueLoop;
+}
+
+DEFINE_HOOK(0x6A57F6, SidebarClass_Init_IO_RecordDiplomacyHouses3, 0x7)
+{
+	enum { ContinueLoop = 0x6A580E, MeetCondition = 0x6A57FF };
+
+	GET(HouseClass*, pHouse, EAX);
+
+	return (pHouse->IsHumanPlayer || HouseClass::CurrentPlayer == HouseClass::Observer) ? MeetCondition : ContinueLoop;
+}